stream/requirements.txt
sseclient
google-cloud-pubsub

